Re-organize codebase.
