--- app/src/main/java/com/example/cactusnotes/LoginActivity.kt
@@ -1,8 +1,0 @@
-package com.example.cactusnotes
-
-import androidx.appcompat.app.AppCompatActivity
-import com.example.cactusnotes.databinding.ActivityLoginBinding
-
-class LoginActivity : AppCompatActivity {
-    lateinit var binding: ActivityLoginBinding
-}
