+++ app/src/main/java/com/example/cactusnotes/api/NotesApi.kt
@@ -1,10 +1,12 @@
-package com.example.cactusnotes
+package com.example.cactusnotes.api
 
+import com.example.cactusnotes.signup.data.RegisterRequest
+import com.example.cactusnotes.signup.data.RegisterResponse
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface SignupAPI {
+interface NotesApi {
     @POST("/auth/local/register")
     fun register(@Body registerRequest: RegisterRequest): Call<RegisterResponse>
 }
