+++ app/src/main/java/com/example/cactusnotes/signup/data/RegisterRequest.kt
@@ -1,4 +1,4 @@
-package com.example.cactusnotes
+package com.example.cactusnotes.signup.data
 
 data class RegisterRequest(
     val email: String, val username: String, val password: String
