+++ app/src/main/java/com/example/cactusnotes/signup/data/RegisterResponse.kt
@@ -1,4 +1,4 @@
-package com.example.cactusnotes
+package com.example.cactusnotes.signup.data
 
 data class RegisterResponse(
     val jwt: String
