+++ app/src/main/java/com/example/cactusnotes/validation/Validator.kt
@@ -1,4 +1,4 @@
-package com.example.cactusnotes
+package com.example.cactusnotes.validation
 
 interface Validator {
     fun validate(field: String): Int?
